+++ CHANGES.md
@@ -2,6 +2,9 @@
 Changes
 =======
 
+- 1.7.3
+    - IMPROVEMENT: optically improve connect button on login dialog
+
 - 1.7.2
     - BUGFIX: optically better grey out meter and sliders on login
     - IMPROVEMENT: optically simplify the login dialog